Extract ReadInstructions helper module from AST parser
This’ll let us use these instruction-reading methods in the preprocessor
too.
